Reorganized tests
Also renamed OpenNet.py, why was it even called that?
