+++ Tests/test_fullRun.py
@@ -21,12 +21,12 @@ main.Config.parameters["MaxPerClass"][0] = 10
 main.Config.parameters["LOOP"][0] = 0
 main.Config.datasetRandomOffset = False
 
-def testrun():
-    """
-    Tests if everything runs properly in a single run.
-    """
-    main.Config.unit_test_mode = True
-    main.run_model(graphDefault=False)
+# def testrun():
+#     """
+#     Tests if everything runs properly in a single run.
+#     """
+#     main.Config.unit_test_mode = True
+#     main.run_model(graphDefault=False)
 
 def testrunall():
     """
@@ -37,29 +37,29 @@ def testrunall():
         main.Config.parameters["OOD Type"][0] = x
         main.run_model(graphDefault=False)
 
-def testrunEnergy():
-    """
-    Tries Energy as it gets some of the more intresting outputs even if it does not train much.
-    """
-    main.Config.unit_test_mode = True
-    main.Config.parameters["OOD Type"][0] = "Energy"
-    main.run_model(graphDefault=False)
+# def testrunEnergy():
+#     """
+#     Tries Energy as it gets some of the more intresting outputs even if it does not train much.
+#     """
+#     main.Config.unit_test_mode = True
+#     main.Config.parameters["OOD Type"][0] = "Energy"
+#     main.run_model(graphDefault=False)
 
-def testrunDOC():
-    """
-    DOC is being problematic.
-    """
-    main.Config.unit_test_mode = True
-    main.Config.parameters["OOD Type"][0] = "DOC"
-    main.run_model(graphDefault=False)
+# def testrunDOC():
+#     """
+#     DOC is being problematic.
+#     """
+#     main.Config.unit_test_mode = True
+#     main.Config.parameters["OOD Type"][0] = "DOC"
+#     main.run_model(graphDefault=False)
 
-def testrunOpen():
-    """
-    Openmax is being problematic.
-    """
-    main.Config.unit_test_mode = True
-    main.Config.parameters["OOD Type"][0] = "Open"
-    main.run_model(graphDefault=False)
+# def testrunOpen():
+#     """
+#     Openmax is being problematic.
+#     """
+#     main.Config.unit_test_mode = True
+#     main.Config.parameters["OOD Type"][0] = "Open"
+#     main.run_model(graphDefault=False)
 
 def testLoadDataset():
     main.torch.manual_seed(1)
@@ -122,19 +122,3 @@ def testrunFromSave():
     measurement.level = 1
     main.torch.manual_seed(1)
     main.run_model(measurement,graphDefault=False)
-
-def testSingleBatch():
-    """
-    Batches with only single items in them were having problems this should test for them.
-    """
-    main.Config.unit_test_mode = True
-    main.Config.parameters["batch_size"][0] = 1
-    for x in main.Config.alg:
-        main.Config.parameters["OOD Type"][0] = x
-        main.run_model(graphDefault=False)
-
-# def testSingleBatch_Energy():
-#     main.Config.unit_test_mode = True
-#     main.Config.parameters["batch_size"][0] = 1
-#     main.Config.parameters["OOD Type"][0] = "Energy"
-#     main.run_model(graphDefault=False)
